improve: Rename RateModelClient => AcrossConfigStoreClient (#39)
* improve: Rename RateModelClient => AcrossConfigStoreClient

* Update ConfigStoreClient.ts

* Update utils.ts

* Update Relayer.IterativeFill.ts

* fix

* lint-fix
