+++ src/clients/index.ts
@@ -1,6 +1,6 @@
 export * from "./HubPoolClient";
 export * from "./SpokePoolClient";
-export * from "./RateModelClient";
+export * from "./ConfigStoreClient";
 export * from "./MultiCallerClient";
 export * from "./ProfitClient";
 export * from "./TokenClient";
