+++ test/fixtures/Dataworker.Fixture.ts
@@ -6,19 +6,11 @@ import {
   BigNumber,
   enableRoutes,
   sampleRateModel,
-  MAX_UINT_VAL,
 } from "../utils";
 import { SignerWithAddress, setupTokensForWallet, getLastBlockTime } from "../utils";
 import { createSpyLogger, winston, deployAndConfigureHubPool, deployConfigStore } from "../utils";
 import * as clients from "../../src/clients";
-import {
-  amountToLp,
-  destinationChainId,
-  originChainId,
-  CHAIN_ID_TEST_LIST,
-  utf8ToHex,
-  repaymentChainId,
-} from "../constants";
+import { amountToLp, destinationChainId, originChainId, CHAIN_ID_TEST_LIST, repaymentChainId } from "../constants";
 
 import { Dataworker } from "../../src/dataworker/Dataworker"; // Tested
 import { TokenClient } from "../../src/clients";
